# Optimize day 1 solution