Local ∆ of `37.33ms => 586.7 μs`.

Maybe surprisingly, this approach to `findTriple` is ~5x faster than a single-scan index-of-sums approach (though that could be an artifact of the implementation / data set).

# Before

```
time                 37.33 ms   (36.67 ms .. 38.77 ms)
                     0.997 R²   (0.993 R² .. 1.000 R²)
mean                 37.42 ms   (37.14 ms .. 38.00 ms)
std dev              834.2 μs   (400.5 μs .. 1.374 ms)
```

# After

```
time                 586.7 μs   (583.5 μs .. 590.0 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 586.7 μs   (584.8 μs .. 589.5 μs)
std dev              7.341 μs   (5.814 μs .. 10.11 μs)
```